✨ Feat: useAmount 구현
AmountContainer 내부의 로직이 길어짐에 따라서 custom hook으로 분리
